requirements.txt
APScheduler==3.6.3
asgiref==3.3.0
astroid==2.4.2
colorama==0.4.4
Django==3.1.3
django-crispy-forms==1.9.2
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumberslite==8.12.12
Pillow==8.0.1
psycopg2==2.8.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
wrapt==1.12.1
